simple_requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
pillow==10.0.0
